refactor: move langgraph/langchain deps to a common place
